% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func.R
\name{accCost2}
\alias{accCost2}
\title{Creates a list with the distances among nodes}
\usage{
accCost2(x, fromCoords)
}
\arguments{
\item{x}{Define this}

\item{fromCoords}{define that}
}
\value{
Creates a list with the distances among nodes
}
\description{
Creates a list with the distances among nodes
}
\author{
Derek Corcoran <derek.corcoran.barrios@gmail.com>

Laurens Geffert <laurensgeffert@gmail.com>
}
