% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_min_cost_flow.R
\name{add_constraints}
\alias{add_constraints}
\title{helper function to add new constraints to an already existsing
list of constraints. Existing list should be length = 3 with elements
- lhs = the left side of the optimization equation (coefficients)
- dir = the condition operation, i.e. '>=', '==', or '<='
- rhs = the constraint value
TODO: add checks for names, length, etc.}
\usage{
add_constraints(constraints, new_lhs, new_dir, new_rhs)
}
\description{
helper function to add new constraints to an already existsing
list of constraints. Existing list should be length = 3 with elements
- lhs = the left side of the optimization equation (coefficients)
- dir = the condition operation, i.e. '>=', '==', or '<='
- rhs = the constraint value
TODO: add checks for names, length, etc.
}
