% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_min_cost_flow.R
\name{data_phillips_problem}
\alias{data_phillips_problem}
\title{helper function to re-create phillips problem in edge list format}
\usage{
data_phillips_problem()
}
\description{
helper function to re-create phillips problem in edge list format
}
