% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispRsal-package.R
\docType{package}
\name{dispRsal}
\alias{dispRsal}
\alias{dispRsal-package}
\title{dispRsal}
\description{
This package implements the minimum cost flow algorithm for biogeographic
dispersal problems
}
\details{

}
\section{Help}{

For more details, see the package vignettes:

\code{browseVignettes(package = "dispRsal")}
}

\examples{
\dontrun{
# Installation
install.packages('devtools')
devtools::install_github(
  repo = '<url>',
  build_vignettes = TRUE
)

library(dispRsal)

# access main help page
?dispRsal

# access vignette (more detailed documentation)
vignette('dispRsal')
}

}
\author{
Laurens Geffert \email{laurensgeffert@gmail.com}
}
