% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func.R
\name{edge_distance_limit}
\alias{edge_distance_limit}
\title{Creates a data frame with the distances among nodes}
\usage{
edge_distance_limit(layer_cost, layers_habitat, Dist)
}
\arguments{
\item{layer_cost}{A raster of cost of each raster cell in the layers_habitat data}

\item{layers_habitat}{A stack with the binary projection of Species Distribution Models where
every layer is a time-slice}

\item{Dist}{Maximum distance in meters that the species can travel between time-slices}
}
\value{
Creates a data frame with the distances among nodes keeping only the
edges that have a distance shorter than Dist
}
\description{
Creates a data frame with the distances among nodes
}
\author{
Derek Corcoran <derek.corcoran.barrios@gmail.com>

Laurens Geffert <laurensgeffert@gmail.com>
}
