% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_min_cost_flow.R
\name{get_dummy_data}
\alias{get_dummy_data}
\title{helper function that creates example raster data
consisting of habitat suitability layers for two time steps
as well as a cost layer}
\usage{
get_dummy_data()
}
\description{
helper function that creates example raster data
consisting of habitat suitability layers for two time steps
as well as a cost layer
}
