% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{PhilCost}
\alias{PhilCost}
\title{A raster with a simulated cost for an hypothetical species}
\format{Raster
\describe{
  \item{PhilCost}{simulated cost layer}
}}
\usage{
PhilCost
}
\description{
This raster has areas with cost Zero (protected areas), unavailable cells
(cities or agricultural areas) and cells with cost.
}
\keyword{datasets}
