% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{Phillips}
\alias{Phillips}
\title{A rasterstack with two timeslices of an hypothetical species}
\format{A rasterstacks of the distribution an hypothetical species:
\describe{
  \item{Phillips}{A stack with two time slices for a Species}
}}
\usage{
Phillips
}
\description{
A dataset containing a stack with two rasters, each represents
a timeslice of the distribution of the species
}
\keyword{datasets}
