% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func.R
\name{solution_linear_old}
\alias{solution_linear_old}
\title{Solves the linear version of Network Flow using the old algorithm}
\usage{
solution_linear_old(layers_habitat, layer_cost, Dist, nchains)
}
\arguments{
\item{layers_habitat}{A stack with the binary projection of Species Distribution Models where
every layer is a time-slice}

\item{layer_cost}{A raster of cost of each raster cell in the layers_habitat data}

\item{Dist}{Maximum distance in meters that the species can travel between time-slices}

\item{nchains}{The minimum number of cells needed for the target species on each time-slice}
}
\value{
A list with a rasterstack with the solution time-slice by time-slice and the global
solution
}
\description{
Solves the linear version of Network Flow using the old algorithm
}
\examples{
# Example 1 based on the Phillips problem
#Load the Species distribution raster
data("Phillips")

# see the dataset
library(raster)
plot(Phillips)

# Load the cost layer

data("PhilCost")

# See the dataset

plot(PhilCost)

# Solve the linear problem with nchains = 2 and a dispersal of 111 km per time-slice

Solution <- solution_linear_old(layers_habitat = Phillips,
layer_cost = PhilCost, Dist = 111000, nchains = 2)

# plot the solution for each time-slice

plot(Solution$FinalStack)

# plot the final solution

plot(Solution$FinalSolution)

#' # Example 2 with hypothetical species
#Load the Species distribution raster
data("BinSpp")

# see the dataset

plot(BinSpp[[1]])

# Load the cost layer

data("Cost")

# See the dataset

plot(Cost)

# Solve the linear problem with nchains = 2 and a dispersal of 111 km per time-slice

Solution <- solution_linear_old(layers_habitat = BinSpp[[1]],
layer_cost = Cost, Dist = 240000, nchains = 5)

# plot the solution for each time-slice

plot(Solution$FinalStack)

# plot the final solution

plot(Solution$FinalSolution)

}
\author{
Derek Corcoran <derek.corcoran.barrios@gmail.com>

Laurens Geffert <laurensgeffert@gmail.com>
}
