% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func.R
\name{add_constraints}
\alias{add_constraints}
\title{helper function to add new constraints to an already existsing
list of constraints. Existing list should be length = 3 with elements
- lhs = the left side of the optimization equation (coefficients)
- dir = the condition operation, i.e. '>=', '==', or '<='
- rhs = the constraint value}
\usage{
add_constraints(constraints, new_lhs, new_dir, new_rhs)
}
\arguments{
\item{constraints}{An constraintsMatrix dataset}

\item{new_lhs}{A new left side of the optimization equation (coefficients)}

\item{new_dir}{A new vector of condition operations, i.e. c('>=', '==', or '<=')}

\item{new_rhs}{A new vector of constraint values}
}
\value{
Adds constraints to a constraints matrix object
}
\description{
helper function to add new constraints to an already existsing
list of constraints. Existing list should be length = 3 with elements
- lhs = the left side of the optimization equation (coefficients)
- dir = the condition operation, i.e. '>=', '==', or '<='
- rhs = the constraint value
}
\author{
Laurens Geffert <laurensgeffert@gmail.com>

Derek Corcoran <derek.corcoran.barrios@gmail.com>
TODO: add checks for names, length, etc.
}
